add a few rules
var:functions: 
Just an array of 60 functions they use to modify canvas contexts. I don't know if it can be of any use.
hook:cellDraw
A hook for drawing a cell when zoomed out. The first parameter is the ID of the arc() method (18) within var:functions. The second one is the ID of a canvas context within var:h.e. The rest are for the arc() method.
var:h
An object containing
e: an array of canvas rendering contexts. Seems to have separate elements for cells, their names and mass values.
images: skins
websockets: contains a websocket for the current server

Hope these aren't completely useless.